Refactor actual copy out to `schematools`
